Add files via upload
This program has a file that has been made to store student names and  marks it is read into. Then the information is loaded into an array then the program determines the high and low marks. An asterisk * is printed next to the highest mark. The program then displays how many people got marks over 50 and above 50.